refactor: move ai chat feature's domain to domain module
