+++ domain/src/main/java/com/emotionstorage/domain/useCase/chat/MarkChatIntroSeenUseCase.kt
@@ -1,0 +1,10 @@
+package com.emotionstorage.domain.useCase.chat
+
+import com.emotionstorage.domain.repo.ChatIntroRepository
+import javax.inject.Inject
+
+class MarkChatIntroSeenUseCase @Inject constructor(
+    private val repository: ChatIntroRepository,
+) {
+    suspend operator fun invoke(value: Boolean) = repository.markIntroSeen(value)
+}
