+++ domain/src/main/java/com/emotionstorage/domain/useCase/chat/ObserveChatMessagesUseCase.kt
@@ -1,7 +1,7 @@
-package com.emotionstorage.ai_chat.domain.usecase.remote
+package com.emotionstorage.domain.useCase.chat
 
-import com.emotionstorage.ai_chat.domain.model.ChatMessage
-import com.emotionstorage.ai_chat.domain.repo.ChatRepository
+import com.emotionstorage.domain.model.ChatMessage
+import com.emotionstorage.domain.repo.ChatRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
