+++ feat/ai-chat/src/main/java/com/emotionstorage/ai_chat/data/dataSource/remote/ChatWSDataSource.kt
@@ -1,6 +1,6 @@
 package com.emotionstorage.ai_chat.data.dataSource.remote
 
-import com.emotionstorage.ai_chat.domain.model.ChatMessage
+import com.emotionstorage.domain.model.ChatMessage
 import kotlinx.coroutines.flow.Flow
 
 interface ChatWSDataSource {
