+++ feat/ai-chat/src/main/java/com/emotionstorage/ai_chat/data/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package com.emotionstorage.ai_chat.data.di
 
 import com.emotionstorage.ai_chat.data.repoImpl.ChatRepositoryImpl
-import com.emotionstorage.ai_chat.domain.repo.ChatRepository
+import com.emotionstorage.domain.repo.ChatRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
