+++ feat/ai-chat/src/main/java/com/emotionstorage/ai_chat/data/repoImpl/AiChatIntroRepositoryImpl.kt
@@ -1,13 +1,13 @@
 package com.emotionstorage.ai_chat.data.repoImpl
 
 import com.emotionstorage.ai_chat.data.dataSource.local.AiChatIntroLocalDataSource
-import com.emotionstorage.ai_chat.domain.repo.AiChatIntroRepository
+import com.emotionstorage.domain.repo.ChatIntroRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class AiChatIntroRepositoryImpl @Inject constructor(
     private val localDataSource: AiChatIntroLocalDataSource,
-) : AiChatIntroRepository {
+) : ChatIntroRepository {
     override fun observeIntroSeen(): Flow<Boolean> = localDataSource.observeIntroSeen()
 
     override suspend fun markIntroSeen(value: Boolean) {
