--- feat/ai-chat/src/main/java/com/emotionstorage/ai_chat/domain/usecase/local/ObserveAiChatIntroSeenUseCase.kt
@@ -1,10 +1,0 @@
-package com.emotionstorage.ai_chat.domain.usecase.local
-
-import com.emotionstorage.ai_chat.domain.repo.AiChatIntroRepository
-import javax.inject.Inject
-
-class ObserveAiChatIntroSeenUseCase @Inject constructor(
-    private val repository: AiChatIntroRepository,
-) {
-    operator fun invoke() = repository.observeIntroSeen()
-}
