+++ feat/ai-chat/src/main/java/com/emotionstorage/ai_chat/presentation/AIChatViewModel.kt
@@ -2,11 +2,11 @@ package com.emotionstorage.ai_chat.presentation
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.emotionstorage.ai_chat.domain.model.ChatMessage
-import com.emotionstorage.ai_chat.domain.usecase.remote.ConnectChatRoomUseCase
-import com.emotionstorage.ai_chat.domain.usecase.remote.DisconnectChatRoomUseCase
-import com.emotionstorage.ai_chat.domain.usecase.remote.ObserveChatMessagesUseCase
-import com.emotionstorage.ai_chat.domain.usecase.remote.SendChatMessageUseCase
+import com.emotionstorage.domain.model.ChatMessage
+import com.emotionstorage.domain.useCase.chat.ConnectChatRoomUseCase
+import com.emotionstorage.domain.useCase.chat.DisconnectChatRoomUseCase
+import com.emotionstorage.domain.useCase.chat.ObserveChatMessagesUseCase
+import com.emotionstorage.domain.useCase.chat.SendChatMessageUseCase
 import com.emotionstorage.domain.common.DataState
 import com.orhanobut.logger.Logger
 import dagger.hilt.android.lifecycle.HiltViewModel
