+++ feat/home/build.gradle.kts
@@ -29,5 +29,4 @@ dependencies {
     implementation(projects.core.common)
     implementation(projects.core.remote)
     implementation(projects.core.ui)
-    implementation(projects.feat.aiChat)
 }
