+++ feat/time-capsule/build.gradle.kts
@@ -27,5 +27,4 @@ dependencies {
     implementation(projects.domain)
     implementation(projects.core.common)
     implementation(projects.core.ui)
-    implementation(projects.feat.aiChat)
 }
